requirements.txt
mysqlclient>=1.3.14
privex-helpers>=2.18.0
python-dotenv
colorama
